Refactor the transfer note and evidence note view details model
