+++ src/EA.Weee.Web.Tests.Unit/ViewModels/EvidenceNoteViewModelTests.cs
@@ -161,24 +161,5 @@
 
             model.DisplayReturnedReason.Should().BeTrue();
         }
-
-        [Theory]
-        [InlineData(NoteStatus.Draft, "Draft evidence note")]
-        [InlineData(NoteStatus.Rejected, "Rejected evidence note")]
-        [InlineData(NoteStatus.Approved, "Approved evidence note")]
-        [InlineData(NoteStatus.Returned, "Returned evidence note")]
-        [InlineData(NoteStatus.Submitted, "Submitted evidence note")]
-        [InlineData(NoteStatus.Void, "")]
-        public void TabName_GivenNoteStatus_ShouldHaveCorrectTabName(NoteStatus status, string expected)
-        {
-            //arrange
-            model.Status = status;
-
-            //act
-            var result = model.TabName;
-
-            //assert
-            result.Should().Be(expected);
-        }
     }
 }
