+++ src/EA.Weee.Web/ViewModels/Shared/EvidenceNoteViewModel.cs
@@ -104,27 +104,5 @@
                 return ManageEvidenceOverviewDisplayOption.ViewAllOtherEvidenceNotes.ToDisplayString();
             }
         }
-
-        public string TabName
-        {
-            get
-            {
-                switch (Status)
-                {
-                    case NoteStatus.Draft:
-                        return "Draft evidence note";
-                    case NoteStatus.Rejected:
-                        return "Rejected evidence note";
-                    case NoteStatus.Approved:
-                        return "Approved evidence note";
-                    case NoteStatus.Returned:
-                        return "Returned evidence note";
-                    case NoteStatus.Submitted:
-                        return "Submitted evidence note";
-                    default:
-                        return string.Empty;
-                }
-            }
-        }
     }
 }
